api: move REST api views to api/views.py
